backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Config / validation
pydantic-settings>=2.5.0
python-dotenv>=1.0.0

# Qwen3.5-Vision — OpenAI-compatible SDK
openai>=1.50.0

# Wan / HTTP calls
httpx>=0.27.0

# Alibaba Cloud OSS
oss2>=2.18.0
